requirements.txt
discord.py
tinydb
flask
gevent
requests
